third_party/blink/web_tests/platform/webview/external/wpt/pointerevents/pointerevent_after_target_removed_interleaved.tentative_mouse-expected.txt

This is a testharness.js-based test.
[FAIL] mouse events from mouse received before/after child removal at pointerdown
  assert_equals: events received expected "mouseover@child,mouseenter@parent,mouseenter@child,(child-removed),mouseover@parent,mousedown@parent,mouseup@parent,click@parent,mousedown@parent,mouseup@parent,click@parent,mouseout@parent,mouseleave@parent" but got "mouseover@child,mouseenter@parent,mouseenter@child,(child-removed),mousedown@parent,mouseover@parent,mouseup@parent,mousedown@parent,mouseup@parent,mouseout@parent,mouseleave@parent"
[FAIL] mouse events from mouse received before/after child removal at pointerup
  assert_equals: events received expected "mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,(child-removed),mouseover@parent,mouseup@parent,click@parent,mousedown@parent,mouseup@parent,click@parent,mouseout@parent,mouseleave@parent" but got "mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,(child-removed),mouseup@parent,mouseover@parent,mousedown@parent,mouseup@parent,mouseout@parent,mouseleave@parent"
Harness: the test ran to completion.
